requirements.txt
requests>=2.32.0
python-dotenv>=1.0.1
azure-identity>=1.19.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
azure-storage-blob>=12.19.0
gunicorn>=21.0.0
asyncpg>=0.29.0
pgvector>=0.2.4
numpy>=1.24.0
openai>=1.50.0
pydantic>=2.9.0
tiktoken>=0.5.0